Remove RouteUtilADM and inline remain code to KnoraResponseRenderer
