+++ webapi/src/main/scala/org/knora/webapi/slice/admin/api/service/ProjectRestService.scala
@@ -101,7 +101,7 @@ final case class ProjectRestService(
            )
     internal <- projectService.createProject(createReq).map(ProjectOperationResponseADM.apply)
     _        <- permissionResponder.createPermissionsForAdminsAndMembersOfNewProject(internal.project.projectIri)
-    external <- format.toExternalADM(internal)
+    external <- format.toExternal(internal)
   } yield external
 
   /**
@@ -121,7 +121,7 @@ final case class ProjectRestService(
       internal <- projectService
                     .updateProject(project, ProjectUpdateRequest(status = Some(Status.Inactive)))
                     .map(ProjectOperationResponseADM.apply)
-      external <- format.toExternalADM(internal)
+      external <- format.toExternal(internal)
     } yield external
 
   /**
@@ -146,7 +146,7 @@ final case class ProjectRestService(
                  .someOrFail(NotFoundException(s"Project '${projectIri.value}' not found."))
     _        <- auth.ensureSystemAdminOrProjectAdmin(user, project)
     internal <- projectService.updateProject(project, updateReq).map(ProjectOperationResponseADM.apply)
-    external <- format.toExternalADM(internal)
+    external <- format.toExternal(internal)
   } yield external
 
   /**
@@ -183,7 +183,7 @@ final case class ProjectRestService(
     project  <- knoraProjectService.findById(id).someOrFail(NotFoundException(s"Project '${getId(id)}' not found."))
     _        <- auth.ensureSystemAdminOrProjectAdmin(user, project)
     internal <- userService.findByProjectMembership(project).map(ProjectMembersGetResponseADM.apply)
-    external <- format.toExternalADM(internal)
+    external <- format.toExternal(internal)
   } yield external
 
   /**
@@ -204,7 +204,7 @@ final case class ProjectRestService(
     project  <- knoraProjectService.findById(id).someOrFail(NotFoundException(s"Project '${getId(id)}' not found."))
     _        <- auth.ensureSystemAdminOrProjectAdmin(user, project)
     internal <- userService.findByProjectAdminMembership(project).map(ProjectAdminMembersGetResponseADM.apply)
-    external <- format.toExternalADM(internal)
+    external <- format.toExternal(internal)
   } yield external
 
   /**
