+++ webapi/src/main/scala/org/knora/webapi/slice/admin/api/service/UserRestService.scala
@@ -55,7 +55,7 @@ final case class UserRestService(
                   .filterOrFail(_.nonEmpty)(NotFoundException("No users found"))
                   .map(_.map(_.filterUserInformation(requestingUser, UserInformationType.Restricted)).sorted)
                   .map(UsersGetResponseADM.apply)
-    external <- format.toExternalADM(internal)
+    external <- format.toExternal(internal)
   } yield external
 
   def deleteUser(requestingUser: User, deleteIri: UserIri): Task[UserResponseADM] = for {
@@ -78,7 +78,7 @@ final case class UserRestService(
       .findUserByIri(userIri)
       .map(_.map(_.groups).getOrElse(Seq.empty))
       .map(UserGroupMembershipsGetResponseADM.apply)
-      .flatMap(format.toExternalADM)
+      .flatMap(format.toExternal)
 
   def createUser(requestingUser: User, userCreateRequest: Requests.UserCreateRequest): Task[UserResponseADM] =
     for {
@@ -92,7 +92,7 @@ final case class UserRestService(
     for {
       kUser    <- getKnoraUserOrNotFound(userIri)
       projects <- projectService.findByIds(kUser.isInProject)
-      external <- format.toExternalADM(UserProjectMembershipsGetResponseADM(projects))
+      external <- format.toExternal(UserProjectMembershipsGetResponseADM(projects))
     } yield external
 
   private def getKnoraUserOrNotFound(userIri: UserIri) =
@@ -102,7 +102,7 @@ final case class UserRestService(
     for {
       kUser    <- getKnoraUserOrNotFound(userIri)
       projects <- projectService.findByIds(kUser.isInProjectAdminGroup)
-      external <- format.toExternalADM(UserProjectAdminMembershipsGetResponseADM(projects))
+      external <- format.toExternal(UserProjectAdminMembershipsGetResponseADM(projects))
     } yield external
 
   def getUserByUsername(requestingUser: User, username: Username): Task[UserResponseADM] = for {
@@ -213,7 +213,7 @@ final case class UserRestService(
 
   private def asExternalUserResponseADM(requestingUser: User, user: User): Task[UserResponseADM] = {
     val userFiltered = user.filterUserInformation(requestingUser, UserInformationType.Restricted)
-    format.toExternalADM(UserResponseADM(userFiltered))
+    format.toExternal(UserResponseADM(userFiltered))
   }
 
   def addUserToProjectAsAdmin(
